pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "textbook2audiobook"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "mss>=9.0.0",
    "Pillow>=10.4.0",
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.scripts]
textbook2audiobook = "main:cli"
